(FACT-3436) Standardize Debian architecture fact
Prior to this commit the os.architecture fact on Debian and
Debian-based distrubitions resolved inconsistently. The fact used
Debian's architecture naming scheme for some platforms (i386 and
amd64) but not others.

This commit updates the os.architecture fact on Debian and
Debian-based distrubitions to consistently use Debian's naming
convention by relying on the output of `dpkg --print-architecture`.